// projects/common/src/lib/interfaces/finding.interface.ts
export interface FindingInterface {
    id: string;
    entity_id: string;
    organization_id: string;
    account_id: string;
    request_id: string;
    service: string;
    platform: string;
    version: string;
    notes: string;
    run_datetime: string;
    created: string;
    updated: string;
    deleted: string;
    payload: string;
}